ECA-7305: Remove unnecessary cast
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@29901 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-common/src/org/ejbca/config/AcmeConfiguration.java
@@ -195,7 +195,7 @@ public class AcmeConfiguration extends UpgradeableDataHashMap implements Seriali
     
     public int getDnsPort() {
         final Integer dnsPort = (Integer) super.data.get(KEY_DNS_PORT);
-        return dnsPort != null ? (int) dnsPort : DNS_SERVER_PORT_DEFAULT;
+        return dnsPort != null ? dnsPort : DNS_SERVER_PORT_DEFAULT;
     }
     
     public void setDnsPort(final int dnsPort) {
